refactor(ui): extract platform/chat/llm/monitoring + relocate 27 types (UI-2 S3)
New domain clients carry their own response types (ex-inline in api.ts):
- platform.ts: OperationsMetrics, BusinessMetrics, TopAPI, ComponentStatus,
  GitOpsStatus, PlatformEvent, ExternalLinks, PlatformStatusResponse,
  ApplicationDiff* (10 types)
- chat.ts: TenantChatSettings, ChatSourceEntry, ChatUsageBySource,
  TokenBudgetStatus, TokenUsageStats, ChatConversationMetrics,
  ModelDistributionEntry, ChatModelDistribution (8 types)
- llm.ts: LlmUsageResponse, LlmTimeseriesPoint, LlmTimeseriesResponse,
  LlmProviderCostEntry, LlmProviderBreakdownResponse, LlmBudgetResponse
  (6 types)
- monitoring.ts: MonitoringTransaction, MonitoringTransactionDetail,
  MonitoringStats (3 types)

api.ts re-exports all 27 types via explicit sub-paths so existing caller
imports (import { X } from 'services/api') keep working.

Size: api.ts 1732 -> 1248 LOC. 2084 tests pass.
